Added Vehicle Monitor Task event flag asserts back in
I removed the asserts initially because they were constantly failing. The issue was the assert logic and the osFlagsWaitAny option was being used with osEventFlagsWait(); this is a mistake. Changing to osFlagsNoClear and manually clearing each flag on pend, using osEventFlagsClear(), fixes this issue.

The directionAlertTimerCallback does, indeed, trigger more than once. It is, however, inconsistent. I'll talk a look at the logic later.

The Vehicle Monitor Task does not yet PASS functional testing, as the logic for the directionAlertTimer being stopped and started seems off. It does, however not cause any other tasks to fail their functional tests. Almost there.